# Improve AutoMixParser performance