Improve AutoMixParser performance

- 目前使用 hashmap 去紀錄每次讀到的值，最後再按照優先順序生成物件，而非解析三次 XML
- 修正 TAGS_MAPPING_TABLE 部分錯誤

測試 7kb的 xml 執行 10 次

測試結果 -> 平均可節省1倍時間

```
 1
 old parser took 35 ms
 new parser took 9 ms
 -------------------------
 2
 old parser took 17 ms
 new parser took 10 ms
 -------------------------
 3
 old parser took 10 ms
 new parser took 3 ms
 -------------------------
 4
 old parser took 6 ms
 new parser took 3 ms
 -------------------------
 5
 old parser took 4 ms
 new parser took 4 ms
 -------------------------
 6
 old parser took 5 ms
 new parser took 2 ms
 -------------------------
 7
 old parser took 6 ms
 new parser took 4 ms
 -------------------------
 8
 old parser took 6 ms
 new parser took 4 ms
 -------------------------
 9
 old parser took 3 ms
 new parser took 1 ms
 -------------------------
 10
 old parser took 2 ms
 new parser took 1 ms
 -------------------------
```